requirements.txt
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
datetime-truncate==1.1.1
Django==4.2.13
django-analytics==0.0.1
django-bower==5.2.0
django-ckeditor-5==0.2.18
django-cleanup==9.0.0
django-geckoboard==2.0.0
django-js-asset==2.2.0
django-memoize==2.3.1
django-multiselectfield==1.0.1
django-nvd3==0.10.1
django-resized==1.0.3
django-taggit==6.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-nvd3==0.16.0
python-slugify==8.0.4
python-telegram-bot==22.1
requests==2.32.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.14.0
urllib3==2.5.0
